requirements.txt
streamlit>=1.36
python-dotenv>=1.0
supabase>=2.5
requests>=2.31
pandas>=2.2
pytz>=2024.1
